+++ frontend/playwright/test-cases/workflows/workflow-release.spec.ts
@@ -49,11 +49,12 @@ ${workflowInputData}
 
   // Add a component to the workflow
   await hetidaDesigner.clickComponentsInNavigation();
+  await hetidaDesigner.searchInNavigation(componentName);
   await hetidaDesigner.clickCategoryInNavigation(componentCategory);
   await hetidaDesigner.dragAndDropItemFromNavigationToFlowchart(
-    componentCategory,
-    `${componentName} (${componentTag})`
+    `${componentName}(${componentTag})`
   );
+  await hetidaDesigner.searchInNavigation(`Test release a workflow ${browserName}`);
 
   // Configure workflow I/O
   await hetidaDesigner.clickIconInToolbar('Configure_IO');
@@ -185,10 +186,7 @@ test.afterEach(async ({ page, hetidaDesigner, browserName }) => {
 
   await hetidaDesigner.clickWorkflowsInNavigation();
   await hetidaDesigner.clickCategoryInNavigation(workflowCategory);
-  await hetidaDesigner.rightClickItemInNavigation(
-    workflowCategory,
-    workflowName
-  );
+  await hetidaDesigner.rightClickItemInNavigation(`${workflowName}(${workflowTag})`);
   await page.locator('.mat-mdc-menu-panel').hover();
 
   if (