+++ frontend/playwright/test-cases/workflows/workflows-list-hover.spec.ts
@@ -4,12 +4,13 @@ test('Hover over workflow, load popover', async ({ page, hetidaDesigner }) => {
   // Arrange
   const categoryName = 'Examples';
   const workflowName = 'Volatility Detection Example';
+  const workflowTag = '1.0.0';
 
   // Act
   await hetidaDesigner.clickWorkflowsInNavigation();
   await hetidaDesigner.clickCategoryInNavigation(categoryName);
 
-  await hetidaDesigner.hoverItemInNavigation(categoryName, workflowName);
+  await hetidaDesigner.hoverItemInNavigation(`${workflowName}(${workflowTag})`);
 
   // Assert
   const popover = page.locator('hd-popover-transformation');