+++ frontend/src/app/app.module.ts
@@ -3,7 +3,12 @@ import {
   provideHttpClient,
   withInterceptorsFromDi
 } from '@angular/common/http';
-import { APP_INITIALIZER, ErrorHandler, NgModule } from '@angular/core';
+import {
+  ErrorHandler,
+  NgModule,
+  inject,
+  provideAppInitializer
+} from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {
   MAT_DIALOG_DATA,
@@ -33,7 +38,10 @@ import {
   WiringTheme
 } from 'hd-wiring';
 import { NgHetidaFlowchartModule } from 'ng-hetida-flowchart';
-import { MonacoEditorModule } from 'ngx-monaco-editor-v2';
+import {
+  MonacoEditorModule,
+  NgxMonacoEditorConfig
+} from 'ngx-monaco-editor-v2';
 import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -93,6 +101,12 @@ const httpLoaderFactory = (configService: ConfigService) => {
   return new StsConfigHttpLoader(authConfig);
 };
 
+const monacoConfig: NgxMonacoEditorConfig = {
+  // TODO: Broken default baseUrl seems fixed in "monaco-editor" version "0.54.0".
+  // https://github.com/microsoft/monaco-editor/issues/4778
+  baseUrl: `${window.location.origin}/assets/monaco/min/vs`
+};
+
 @NgModule({
   declarations: [
     AppComponent,
@@ -133,9 +147,7 @@ const httpLoaderFactory = (configService: ConfigService) => {
     OwlMomentDateTimeModule,
     MaterialModule,
     HdWiringModule,
-    MonacoEditorModule.forRoot({
-      baseUrl: './assets'
-    }), // use forRoot() in main app module only.
+    MonacoEditorModule.forRoot(monacoConfig), // use forRoot() in main app module only.
     StoreModule.forRoot(appReducers, {
       runtimeChecks: {
         strictStateImmutability: false,
@@ -178,16 +190,14 @@ const httpLoaderFactory = (configService: ConfigService) => {
     },
     { provide: MatDialogRef, useValue: {} },
     ConfigService,
-    {
-      provide: APP_INITIALIZER,
-      useFactory: (appConfig: ConfigService) => {
+    provideAppInitializer(() => {
+      const initializerFn = ((appConfig: ConfigService) => {
         return async () => {
           await appConfig.loadConfig();
         };
-      },
-      multi: true,
-      deps: [ConfigService]
-    },
+      })(inject(ConfigService));
+      return initializerFn();
+    }),
     {
       provide: HD_WIRING_CONFIG,
       useFactory: (themeService: ThemeService): HdWiringConfig => {