+++ frontend/src/app/components/auth-callback/auth-callback.component.ts
@@ -3,7 +3,8 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'hd-auth-callback',
   templateUrl: './auth-callback.component.html',
-  styleUrls: ['./auth-callback.component.scss']
+  styleUrls: ['./auth-callback.component.scss'],
+  standalone: false
 })
 export class AuthCallbackComponent {
   constructor() {}