+++ frontend/src/app/components/component-io-dialog/component-io-dialog.component.ts
@@ -32,7 +32,8 @@ export interface ComponentIoDialogData {
 @Component({
   selector: 'hd-component-new-revision-modal',
   templateUrl: 'component-io-dialog.component.html',
-  styleUrls: ['component-io-dialog.component.scss']
+  styleUrls: ['component-io-dialog.component.scss'],
+  standalone: false
 })
 export class ComponentIODialogComponent implements AfterViewInit {
   componentTransformation: ComponentTransformation;