+++ frontend/src/app/components/copy-transformation-dialog/copy-transformation-dialog.component.ts
@@ -17,7 +17,8 @@ import { ConfigService } from '../../service/configuration/config.service';
 @Component({
   selector: 'hd-copy-transformation-dialog',
   templateUrl: './copy-transformation-dialog.component.html',
-  styleUrls: ['./copy-transformation-dialog.component.scss']
+  styleUrls: ['./copy-transformation-dialog.component.scss'],
+  standalone: false
 })
 export class CopyTransformationDialogComponent implements OnInit {
   private apiEndpoint: string;