+++ frontend/src/app/components/home/home.component.ts
@@ -7,7 +7,8 @@ import { ThemeService } from './../../service/theme/theme.service';
 @Component({
   selector: 'hd-home',
   templateUrl: './home.component.html',
-  styleUrls: ['./home.component.scss']
+  styleUrls: ['./home.component.scss'],
+  standalone: false
 })
 export class HomeComponent {
   constructor(