+++ frontend/src/app/components/popover-transformation/popover-transformation.component.ts
@@ -23,7 +23,8 @@ import { selectTransformationById } from '../../store/transformation/transformat
 @Component({
   selector: 'hd-popover-transformation',
   templateUrl: './popover-transformation.component.html',
-  styleUrls: ['./popover-transformation.component.scss']
+  styleUrls: ['./popover-transformation.component.scss'],
+  standalone: false
 })
 export class PopoverTransformationComponent implements OnInit {
   @HostBinding('style.top')